flask_app/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
gunicorn==23.0.0
packaging==24.2
python-dateutil==2.8.2

